refactor(assistant): merge RoutineProcess into AgentProcess
AgentProcess now detects whether its input is a JSON-encoded
RoutineRunInput and, if so, executes the routine inline (one-shot),
emitting a RoutineRunOutput and calling ctx.succeed() to self-terminate.
Plain string inputs continue through the existing alarm-driven agent path.

routine-process.ts is deleted; all exports are re-homed in agent-process.ts.
