+++ packages/core/assistant/src/service/agent-process.ts
@@ -4,25 +4,30 @@
 
 import * as Tool from '@effect/ai/Tool';
 import * as Toolkit from '@effect/ai/Toolkit';
+import * as Array from 'effect/Array';
 import * as Cause from 'effect/Cause';
+import * as Deferred from 'effect/Deferred';
 import * as Duration from 'effect/Duration';
 import * as Effect from 'effect/Effect';
 import * as Exit from 'effect/Exit';
+import * as Function from 'effect/Function';
 import * as Layer from 'effect/Layer';
 import * as Match from 'effect/Match';
 import * as Option from 'effect/Option';
 import * as Schema from 'effect/Schema';
 
 import { AiService, OpaqueToolkit, type ModelName } from '@dxos/ai';
+import { Routine, Template } from '@dxos/compute';
 import { Trace } from '@dxos/compute';
 import { Operation, OperationRegistry } from '@dxos/compute';
-import { Database, DXN, Feed, Obj } from '@dxos/echo';
+import { Database, DXN, Feed, Obj, Ref } from '@dxos/echo';
 import { acquireReleaseResource } from '@dxos/effect';
 import { Process, ProcessManager, StorageService } from '@dxos/functions-runtime';
 import { log } from '@dxos/log';
 import { trim } from '@dxos/util';
 
 import { type McpServerConfig, AiSession } from '../conversation';
+import { RoutineError } from '../errors';
 import { getOperationFromTool, makeToolExecutionService, makeToolResolverFromOperations } from '../functions';
 import { AgentRequestBegin, AgentRequestEnd } from '../tracing';
 
@@ -38,8 +43,27 @@ interface AgentProcessOptions {
 
 export const AGENT_PROCESS_KEY = 'org.dxos.testing.process.agent';
 
+export const RoutineRunInput = Schema.Struct({
+  routineDxn: Schema.String,
+  feedDxn: Schema.String,
+  input: Schema.optional(Schema.Any),
+  systemInstructions: Schema.optional(Schema.String),
+  model: Schema.optional(Schema.String),
+});
+export type RoutineRunInput = Schema.Schema.Type<typeof RoutineRunInput>;
+
+const RoutineRunOutput = Schema.Union(
+  Schema.TaggedStruct('success', { value: Schema.Any }),
+  Schema.TaggedStruct('failure', {
+    message: Schema.String,
+    description: Schema.optional(Schema.String),
+  }),
+);
+export type RoutineRunOutput = Schema.Schema.Type<typeof RoutineRunOutput>;
+
 /**
  * Hosts a persistent, suspendible AiAgent that can process a number of prompts.
+ * Also handles one-shot routine execution when the input is a JSON-encoded {@link RoutineRunInput}.
  * The process target is a queue DXN string.
  */
 export const AgentProcess = (options: AgentProcessOptions) =>
@@ -47,7 +71,7 @@ export const AgentProcess = (options: AgentProcessOptions) =>
     {
       key: AGENT_PROCESS_KEY,
       input: Schema.String,
-      output: Schema.Void,
+      output: RoutineRunOutput,
       services: [
         Database.Service,
         OpaqueToolkit.OpaqueToolkitProvider,
@@ -74,9 +98,19 @@ export const AgentProcess = (options: AgentProcessOptions) =>
         yield* toolCallManager.load();
 
         return {
-          onInput: Effect.fnUntraced(function* (prompt: string) {
-            log('agent onInput received', { promptLength: prompt.length, backlog: inputQueue.length });
-            inputQueue.push({ _tag: 'prompt', content: prompt });
+          onInput: Effect.fnUntraced(function* (input: string) {
+            // Detect routine mode: try to decode as RoutineRunInput.
+            const routineInput = yield* Schema.decodeUnknown(Schema.parseJson(RoutineRunInput))(input).pipe(
+              Effect.option,
+            );
+            if (Option.isSome(routineInput)) {
+              yield* runRoutineInput(ctx, session, feed, routineInput.value, options);
+              return;
+            }
+
+            // Agent mode: queue prompt for alarm-driven processing.
+            log('agent onInput received', { promptLength: input.length, backlog: inputQueue.length });
+            inputQueue.push({ _tag: 'prompt', content: input });
             log('agent onInput persisting queue', { depth: inputQueue.length });
             yield* AgentEventsKey.set(inputQueue);
             log('agent onInput persisted', { depth: inputQueue.length });
@@ -179,6 +213,204 @@ export const AgentProcess = (options: AgentProcessOptions) =>
         };
       }),
   );
+
+//
+// Routine execution.
+//
+
+type RoutineCtx = Process.ProcessContext<string, RoutineRunOutput>;
+
+const runRoutineInput = (
+  ctx: RoutineCtx,
+  session: AiSession,
+  feed: Feed.Feed,
+  runInput: RoutineRunInput,
+  options: AgentProcessOptions,
+) => {
+  const run = Effect.gen(function* () {
+    log.info('routine input: received', { routineDxn: runInput.routineDxn });
+
+    const routine = yield* Database.resolve(DXN.parse(runInput.routineDxn), Routine.Routine).pipe(
+      Effect.mapError((err) => new RoutineError('Failed to resolve routine.', { description: String(err) })),
+    );
+
+    const blueprints = yield* Function.pipe(
+      routine.blueprints,
+      Effect.forEach(Database.loadOption),
+      Effect.map(Array.filter(Option.isSome)),
+      Effect.map(Array.map((opt) => opt.value)),
+    );
+
+    const objects = yield* Function.pipe(
+      routine.context,
+      Effect.forEach(Database.loadOption),
+      Effect.map(Array.filter(Option.isSome)),
+      Effect.map(Array.map((opt) => opt.value)),
+    );
+
+    const promptInstructions = yield* Database.load(routine.instructions.source).pipe(
+      Effect.mapError((err) => new RoutineError('Failed to load routine instructions.', { description: String(err) })),
+    );
+    let promptText = Template.process(promptInstructions.content, runInput.input);
+    if (runInput.input !== undefined) {
+      promptText += `\n<input>${JSON.stringify(runInput.input)}</input>`;
+    }
+
+    let systemText = trim`
+      You are an agent running in the non-interactive mode.
+      The user is unable to see what you are doing, and cannot answer any questions.
+      Do not ask questions.
+      Complete the task before you, and at the end call [completeJob] with the output.
+      If you are unable to complete the task, call [completeJob] with the failure reason.
+      If no output is required, call [completeJob] with an empty object: {}
+      Do not stop until you call [completeJob].
+    `;
+    if (runInput.systemInstructions) {
+      systemText += `\n\n${runInput.systemInstructions}`;
+    }
+
+    const modelLayer = AiService.model((runInput.model as ModelName) ?? '@anthropic/claude-opus-4-6');
+
+    yield* Effect.promise(() =>
+      session.context.bind({
+        blueprints: blueprints.map((blueprint) => Ref.make(blueprint)),
+        objects: objects.map((object) => Ref.make(object as Obj.Unknown)),
+      }),
+    );
+
+    const resultSink = yield* Deferred.make<unknown, RoutineError>();
+    const completeJobToolkit = makeCompleteJobToolkit({ resultSink });
+
+    const runRequest = (prompt: string) =>
+      session
+        .createRequest({
+          prompt,
+          system: systemText,
+          toolkit: completeJobToolkit,
+          mcpServers: options.getMcpServers?.(),
+        })
+        .pipe(
+          Effect.provide(
+            Layer.mergeAll(modelLayer, RoutineToolExecutionService({ feed }), makeToolResolverFromOperations()),
+          ),
+        );
+
+    yield* runRequest(promptText);
+
+    const pollResult = yield* Deferred.poll(resultSink);
+
+    if (Option.isNone(pollResult)) {
+      yield* runRequest('You must signal task completion by calling [completeJob] with the output or failure reason.');
+      const retryResult = yield* Deferred.poll(resultSink);
+      if (Option.isNone(retryResult)) {
+        ctx.submitOutput({ _tag: 'failure', message: 'Agent did not signal task completion.' });
+        ctx.succeed();
+        return;
+      }
+      yield* emitRoutineResult(ctx, retryResult.value);
+      return;
+    }
+
+    yield* emitRoutineResult(ctx, pollResult.value);
+  });
+
+  return run.pipe(
+    Effect.mapError((err) => (err instanceof RoutineError ? err : new RoutineError(String(err)))),
+    Effect.catchAll((err) => emitRoutineResult(ctx, Effect.fail(err))),
+    Effect.catchAllCause((cause) =>
+      emitRoutineResult(
+        ctx,
+        Effect.fail(new RoutineError('Routine process failed.', { description: Cause.pretty(cause) })),
+      ),
+    ),
+    Effect.scoped,
+  );
+};
+
+const emitRoutineResult = (
+  ctx: RoutineCtx,
+  resultEffect: Effect.Effect<unknown, RoutineError, never>,
+): Effect.Effect<void> =>
+  resultEffect.pipe(
+    Effect.matchEffect({
+      onFailure: (err) =>
+        Effect.sync(() => {
+          ctx.submitOutput({
+            _tag: 'failure',
+            message: err.message,
+            description: err.context?.description as string | undefined,
+          });
+          ctx.succeed();
+        }),
+      onSuccess: (value) =>
+        Effect.sync(() => {
+          ctx.submitOutput({ _tag: 'success', value });
+          ctx.succeed();
+        }),
+    }),
+  );
+
+const makeCompleteJobToolkit = (options: { resultSink: Deferred.Deferred<unknown, RoutineError> }) => {
+  class CompleteJobToolkit extends Toolkit.make(
+    Tool.make('completeJob', {
+      parameters: {
+        success: Schema.optional(Schema.Any),
+        failure: Schema.optional(
+          Schema.Struct({
+            message: Schema.String.annotations({ description: 'Short message describing the error.' }),
+            description: Schema.optional(Schema.String).annotations({
+              description: 'Optional longer message describing in detail what went wrong',
+            }),
+          }),
+        ),
+      },
+    }),
+  ) {}
+
+  const layer = CompleteJobToolkit.toLayer({
+    completeJob: Effect.fnUntraced(function* (result) {
+      if (result.failure) {
+        yield* Deferred.fail(
+          options.resultSink,
+          new RoutineError(result.failure.message, { description: result.failure.description }),
+        );
+      } else {
+        yield* Deferred.succeed(options.resultSink, result.success);
+      }
+    }),
+  });
+
+  return OpaqueToolkit.make(CompleteJobToolkit, layer);
+};
+
+interface RoutineToolExecutionOptions {
+  feed: Feed.Feed;
+}
+
+const RoutineToolExecutionService = ({ feed }: RoutineToolExecutionOptions) =>
+  Layer.unwrapEffect(
+    Effect.gen(function* () {
+      const operationInvoker = yield* Operation.Service;
+      return makeToolExecutionService({
+        invoke: (tool, input) =>
+          Effect.gen(function* () {
+            const operationDef = getOperationFromTool(tool).pipe(Option.getOrThrow);
+            log('invoking operation', { operationDef, input });
+            const result = yield* operationInvoker
+              .invoke(operationDef, input, {
+                conversation: Obj.getDXN(feed).toString(),
+              })
+              .pipe(Effect.orDie);
+            log('result', { result });
+            return result;
+          }),
+      });
+    }),
+  );
+
+//
+// Agent-mode helpers.
+//
 
 interface ToolExecutionServiceOptions {
   /**
