+++ packages/core/assistant/src/service/index.ts
@@ -4,5 +4,4 @@
 
 export * as AgentService from './AgentService';
 export { AGENT_PROCESS_KEY } from './agent-process';
-export { ROUTINE_PROCESS_KEY } from './routine-process';
-export type { RoutineRunInput, RoutineRunOutput } from './routine-process';
+export type { RoutineRunInput, RoutineRunOutput } from './agent-process';
